[test refactor] "Bless" (i.e. undeprecate) splat-friendly API.
