+++ test/logger-legacy.test.js
@@ -110,98 +110,6 @@ describe('Logger (legacy API)', function () {
     })
   );
 
-  it('.log(level, message)', function (done) {
-    var logger = helpers.createLogger(function (info) {
-      assume(info).is.an('object');
-      assume(info.level).equals('info');
-      assume(info.message).equals('Some super awesome log message');
-      assume(info[MESSAGE]).is.a('string');
-      done();
-    });
-
-    logger.log('info', 'Some super awesome log message')
-  });
-
-  it(`.log(level, undefined) creates info with { message: undefined }`, function (done) {
-    const logger = helpers.createLogger(function (info) {
-      assume(info.message).equals(undefined);
-      done();
-    });
-
-    logger.log('info', undefined);
-  });
-
-  it(`.log(level, null) creates info with { message: null }`, function (done) {
-    const logger = helpers.createLogger(function (info) {
-      assume(info.message).equals(null);
-      done();
-    });
-
-    logger.log('info', null);
-  });
-
-  it(`.log(level, new Error()) uses Error instance as info`, function (done) {
-    const err = new Error('test');
-    const logger = helpers.createLogger(function (info) {
-      assume(info).instanceOf(Error);
-      assume(info).equals(err);
-      done();
-    });
-
-    logger.log('info', err);
-  });
-
-  it('.log(level, message, meta)', function (done) {
-    var meta = { one: 2 };
-    var logger = helpers.createLogger(function (info) {
-      assume(info).is.an('object');
-      assume(info.level).equals('info');
-      assume(info.message).equals('Some super awesome log message');
-      assume(info.one).equals(2);
-      assume(info[MESSAGE]).is.a('string');
-      done();
-    });
-
-    logger.log('info', 'Some super awesome log message', meta);
-  });
-
-  it('.log(level, formatStr, ...splat)', function (done) {
-    const format = winston.format.combine(
-      winston.format.splat(),
-      winston.format.printf(info => `${info.level}: ${info.message}`)
-    );
-
-    var logger = helpers.createLogger(function (info) {
-      assume(info).is.an('object');
-      assume(info.level).equals('info');
-      assume(info.message).equals('100% such wow {"much":"javascript"}');
-      assume(info.splat).deep.equals([100, 'wow', { much: 'javascript' }]);
-      assume(info[MESSAGE]).equals('info: 100% such wow {"much":"javascript"}');
-      done();
-    }, format);
-
-    logger.log('info', '%d%% such %s %j', 100, 'wow', { much: 'javascript' });
-  });
-
-  it('.log(level, formatStr, ...splat, meta)', function (done) {
-    const format = winston.format.combine(
-      winston.format.splat(),
-      winston.format.printf(info => `${info.level}: ${info.message} ${JSON.stringify(info.meta)}`)
-    );
-
-    var logger = helpers.createLogger(function (info) {
-      assume(info).is.an('object');
-      assume(info.level).equals('info');
-      assume(info.message).equals('100% such wow {"much":"javascript"}');
-      assume(info.splat).deep.equals([100, 'wow', { much: 'javascript' }]);
-      assume(info.meta).deep.equals({ thisIsMeta: true });
-      assume(info[MESSAGE]).equals('info: 100% such wow {"much":"javascript"} {"thisIsMeta":true}');
-      done();
-    }, format);
-
-    logger.log('info', '%d%% such %s %j', 100, 'wow', { much: 'javascript' }, { thisIsMeta: true });
-  });
-
   it('.cli() throws', function () {
     var logger = winston.createLogger();
     assume(logger.cli).throws(/Use a custom/);
